Moved proposed app instance info; added doc for context
